[Experimental Feature] Redis DB for Control / History
This is an experimental branch of PiFire that uses the Redis DB for messaging between the flask front end and the control backend script.  The intent of using Redis DB is to improve overall performance and avoid collisions when multiple scripts are accessing data at the same time (would sometimes cause collisions with the old file database method).  

A few advantages of using Redis instead of the file system:

1. Redis is in-memory and thus is very fast
2. Since Redis is in-memory, writes to the db are not written to the SD-card with any frequency, so it extend SD-Card life.  
3. It's lightweight and easy to use, and readily available (simple installation)
4. The python module is very simple to install and simple to use

As a result, I've upped the refresh rate of the display, the dashboard, and the history pages.  I might even up the refresh rate of the temperature samples in the future.  

WARNING: This hasn't been fully tested on running PiFire hardware and is for experimental purposes only.  Use at your own risk. 

Further refinement of this implementation will be coming in future releases.  

If you are installing this over the top of your current PiFire instance, you will need to install both the redis-server and the redis-python module.  

In the terminal, first install the redis-server: 
$ sudo apt update
$ sudo apt install redis-server

You might need to start the server:
$ sudo service redis-server start

Then install the redis python module:
$ sudo pip3 install redis 

Then you should be able to reboot and run from there.