Simplified Storing Control JSON
Why was I being soooo hard on myself.  Storing JSON in Redis is simple when you smash it into a string first.  Why didn't I think of that before!  Anyway, it's now way simpler of an implementation.  

BTW - Now tested on real hardware and I'm happy to report it's working great.